refactor(importers): isolate Junction canonical coverage
